requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.2
future==0.18.3
macaddress==2.0.2
netmiko==4.2.0
ntc-templates==3.5.0
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.0
PyYAML==6.0.1
scp==0.14.5
six==1.16.0
textfsm==1.1.3
